+++ dart/geodata/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.7.0
+
+- BETA version 0.7.0 with breaking changes
+- updated dependency 0.7.1 on [attributes](https://pub.dev/packages/attributes)
+- [Official Dart lint rules applied with recommend set #32](https://github.com/navibyte/geospatial/issues/32)
+
 ## 0.6.0
 
 - BETA version 0.6.0 with minor breaking changes